requirements.txt
asgiref==3.8.1
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
djangorestframework==3.15.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pypdfium2==4.28.0
sqlparse==0.4.4
tzdata==2024.1
